2.0.0 - "has-a" / callback interface for Proactor (#344)
Proactor now has a "callback interface" and is expected to be used in a "has-a" relationship with the main actor, rather than in an "is-a" relationship where the main actor inherits from Proactor. This lowers the understanding of Proactor internals required to use Proactor effectively. The expected interaction method is for user code to implement a PrimeActor subclass which will receive messages via the Proactor and an App subclass which sets up the Proactor and the (optional) PrimeActor.

Also adds WebEventListener and Proactor interface functions await_processing() and wait_for_processing_threadsafe().