PrimeActor gets *App* not Proactor (#393)
The PrimeActor now receives the App, not the Proactor. In all existing cases this should be functionally the same, but this allows us to add app-not-proactor related things to the AppInterface later, this change will prevent surprising entanglement. This was the intended design (all other Actor's receive the App, and for most (all?) functions in AppInterface that were previously in ServiceInterface, the App is just a pass through to the Proactor. This resolved issue #383.